Set up the initial cargo-lit command
This creates a very rudimentary program that will turn basic markdown in
your src dir into equivalent rust code. It then will invoke the command
that you would expect cargo to run on the src code that gets output.
This does not create the most ideal cli tool experience nor does it
support much in terms of markdown but it's a start. It also includes
a test crate to see that it works as expected.